+++ benchmarks/src/main/javaTemplates/net/openhft/collect/research/hash/AddRemoveWithTombstonesBenchmarks.java
@@ -45,9 +45,9 @@ public class AddRemoveWithTombstonesBenchmarks {
     static final double LOAD_FACTOR = parseDouble(System.getProperty("loadFactor", "0.5"));
     static final double REHASH_LOAD = parseDouble(System.getProperty("rehashLoad", "0.75"));
 
-    static final int D_HASH_CAPACITY = DHashCapacities.bestCapacity(SIZE, LOAD_FACTOR, 0);
+    static final int D_HASH_CAPACITY = DHashCapacities.bestCapacity((long) SIZE, LOAD_FACTOR, 0);
     static final int Q_HASH_CAPACITY =
-            QHashCapacities.getIntCapacity(((int)(SIZE / LOAD_FACTOR)) + 1, 0);
+            QHashCapacities.getIntCapacity(((int)((double) SIZE / LOAD_FACTOR)) + 1, 0);
 
     static int addRemovesToRehashOnce(double targetLoad, double rehashLoad, int capacity) {
         // Expected theoretical number of add-remove pairs to reach rehash load
